// web/src/gen/model/pinDefinition.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Automation.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { PinCardinality } from "./pinCardinality";
import type { PinKind } from "./pinKind";
import type { PinPrimitiveType } from "./pinPrimitiveType";

export interface PinDefinition {
  id?: string;
  label?: string;
  kind?: PinKind;
  primitiveType?: PinPrimitiveType;
  cardinality?: PinCardinality;
  isRequired?: boolean;
  defaultValue?: unknown;
  /** @nullable */
  metadata?: string | null;
  /** @nullable */
  entityTarget?: string | null;
  /** @nullable */
  allowedExtensions?: string | null;
}